make keychain API public
